+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Added `IntCollector/LongCollector/DoubleCollector.averaging` primitive collectors.
 * Added `IntCollector/LongCollector/DoubleCollector.andThen` default methods to replace `collectingAndThen`.
 * Added `StreamEx.toFlatCollection` and `StreamEx.toFlatList` terminal operations.
+* Added `StreamEx.ofSubLists(list, length, shift)` static method.
 * Methods `IntCollector/LongCollector/DoubleCollector.collectingAndThen` are declared as deprecated and may be removed in future releases!
 
 ### 0.3.6